requirements.txt
torch
torchvision
segmentation-models-pytorch
albumentations
opencv-python
numpy
tqdm
pyyaml